feat: Add some tools to work with gettext
redo python file parser

Issue #16
