+++ tests/lang/__init__.py
@@ -17,6 +17,7 @@
 #
 
 __all__ = [
+    "domain",
     "lang",
     "parse",
     "pyfile"
